Fog and clouds optimized, tone optimized
The darker tone of this game has been revitalized. Fog is now thicker, clouds are more realistic, and now we have a working sunlight with a changeable angle!
Still need to make sure the sun in the sky(box) moves along with it tho, and also fix some minor issues as it gets darker!